+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { LoginModule } from './login/login.module';
 import { RegisterModule } from './register/register.module';
-import { GetuserModule } from './getuser/getuser.module';
+import { GetuserModule } from './getuser/user.module';
 
 @Module({
   imports: [LoginModule, RegisterModule, GetuserModule],
