+++ src/getuser/user.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import {userService } from './user.service';
+import { userController } from './user.controller';
+import { PrismaService } from 'src/prisma.service';
+@Module({
+  controllers: [userController],
+  providers: [userService, PrismaService],
+})
+export class GetuserModule {}
